Refactor lexer.c and corrected norme error in main.c error.c
