+++ src/error.c
@@ -6,7 +6,7 @@
 /*   By: eguefif <eguefif@student.42quebec.>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/14 13:56:24 by eguefif           #+#    #+#             */
-/*   Updated: 2023/11/17 13:11:19 by eguefif          ###   ########.fr       */
+/*   Updated: 2023/11/17 14:03:17 by eguefif          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,5 +21,5 @@ void	ft_error_message(char *s, int error_type)
 {
 	if (error_type == SYNTAX_ERROR)
 		ft_dprintf(2, "%s: syntax error near unexpected token \'%s\'\n",
-			   PROG_NAME, s);
+			PROG_NAME, s);
 }
