+++ src/parser.c
@@ -6,7 +6,7 @@
 /*   By: eguefif <eguefif@student.42quebec.>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/13 12:10:46 by eguefif           #+#    #+#             */
-/*   Updated: 2023/11/17 13:45:36 by eguefif          ###   ########.fr       */
+/*   Updated: 2023/11/17 14:03:20 by eguefif          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
